requirements.txt
streamlit
python-dotenv
requests
openai
pypdf
python-docx
beautifulsoup4
validators
psycopg2-binary
fastapi
uvicorn
python-multipart
